refactor(plugins): move AppRootPage into a /components folder
