+++ public/app/features/plugins/tests/AppRootPage.test.tsx
@@ -1,6 +1,6 @@
 import { act, render, screen } from '@testing-library/react';
 import React, { Component } from 'react';
-import AppRootPage from '../AppRootPage';
+import AppRootPage from '../components/AppRootPage';
 import { getPluginSettings } from '../PluginSettingsCache';
 import { importAppPlugin } from '../plugin_loader';
 import { getMockPlugin } from '../__mocks__/pluginMocks';
